Tweak/rename FormForUsingPortableClipboard
